Moving user/profile and user/dashboard pages to server components
