+++ src/app/user/dashboard/loading.tsx
@@ -1,0 +1,5 @@
+import { LoadingSpinner } from "@/components/ui/loading-spinner";
+
+export default function Loading() {
+  return <LoadingSpinner />;
+}
